move cucumber examples out of subdirectory
