+++ selenium-cucumber-examples/src/test/resources/RemoveItem.feature
